auth-service/requirements.txt
# requirements.txt
Flask
Flask-Cors
PyJWT
requests
PyMySQL
gunicorn
cryptography